+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - missing files encountered during searching will log warnings gracefully instead of harshly
+- handle newline characters in the TUI (#98)
 
 
 ## [3.5.3] - 2022-11-16